fixed error: attribute type Relation incompatiable

+++ database/migrations/2021_03_12_074638_create_product_images_table.php
@@ -15,7 +15,7 @@ class CreateProductImagesTable extends Migration
     {
         Schema::create('product_images', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('product_id')->index();
+            $table->unsignedBigInteger('product_id')->index();
             $table->string('thumbnail')->nullable();
             $table->string('full')->nullable();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
