+++ database/migrations/2021_03_12_080940_create_attribute_value_product_attribute_table.php
@@ -15,9 +15,9 @@ class CreateAttributeValueProductAttributeTable extends Migration
     {
         Schema::create('attribute_value_product_attribute', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('attribute_value_id');
+            $table->unsignedBigInteger('attribute_value_id');
             $table->foreign('attribute_value_id')->references('id')->on('attribute_values');
-            $table->unsignedInteger('product_attribute_id');
+            $table->unsignedBigInteger('product_attribute_id');
             $table->foreign('product_attribute_id')->references('id')->on('product_attributes');
             $table->timestamps();
         });
